+++ apiserver/application/mock_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/juju/juju/state"
 	statestorage "github.com/juju/juju/state/storage"
 	coretesting "github.com/juju/juju/testing"
-	"gopkg.in/macaroon.v1"
 )
 
 type mockEnviron struct {
