+++ apiserver/crossmodelrelations/crossmodelrelations_test.go
@@ -1,0 +1,139 @@
+// Copyright 2017 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package crossmodelrelations_test
+
+import (
+	"github.com/juju/testing"
+	jc "github.com/juju/testing/checkers"
+	gc "gopkg.in/check.v1"
+	"gopkg.in/juju/charm.v6-unstable"
+	"gopkg.in/juju/names.v2"
+
+	"github.com/juju/juju/apiserver/common"
+	"github.com/juju/juju/apiserver/crossmodelrelations"
+	"github.com/juju/juju/apiserver/params"
+	apiservertesting "github.com/juju/juju/apiserver/testing"
+	"github.com/juju/juju/core/crossmodel"
+	"github.com/juju/juju/state"
+	coretesting "github.com/juju/juju/testing"
+)
+
+var _ = gc.Suite(&remoteRelationsSuite{})
+
+type remoteRelationsSuite struct {
+	coretesting.BaseSuite
+
+	resources  *common.Resources
+	authorizer *apiservertesting.FakeAuthorizer
+	st         *mockState
+	api        *crossmodelrelations.CrossModelRelationsAPI
+}
+
+func (s *remoteRelationsSuite) SetUpTest(c *gc.C) {
+	s.BaseSuite.SetUpTest(c)
+
+	s.resources = common.NewResources()
+	s.AddCleanup(func(_ *gc.C) { s.resources.StopAll() })
+
+	s.authorizer = &apiservertesting.FakeAuthorizer{
+		Tag:        names.NewMachineTag("0"),
+		Controller: true,
+	}
+
+	s.st = newMockState()
+	api, err := crossmodelrelations.NewCrossModelRelationsAPI(s.st, s.resources, s.authorizer)
+	c.Assert(err, jc.ErrorIsNil)
+	s.api = api
+}
+
+func (s *remoteRelationsSuite) TestPublishLocalRelationsChange(c *gc.C) {
+	s.st.remoteApplications["db2"] = &mockRemoteApplication{}
+	s.st.remoteEntities[names.NewApplicationTag("db2")] = "token-db2"
+	rel := newMockRelation(1)
+	ru1 := newMockRelationUnit()
+	ru2 := newMockRelationUnit()
+	rel.units["db2/1"] = ru1
+	rel.units["db2/2"] = ru2
+	s.st.relations["db2:db django:db"] = rel
+	s.st.remoteEntities[names.NewRelationTag("db2:db django:db")] = "token-db2:db django:db"
+	results, err := s.api.PublishLocalRelationChange(params.RemoteRelationsChanges{
+		Changes: []params.RemoteRelationChangeEvent{
+			{
+				Life: params.Alive,
+				ApplicationId: params.RemoteEntityId{
+					ModelUUID: "uuid",
+					Token:     "token-db2"},
+				RelationId: params.RemoteEntityId{
+					ModelUUID: "uuid",
+					Token:     "token-db2:db django:db"},
+				ChangedUnits: []params.RemoteRelationUnitChange{{
+					UnitId:   1,
+					Settings: map[string]interface{}{"foo": "bar"},
+				}},
+				DepartedUnits: []int{2},
+			},
+		},
+	})
+	c.Assert(err, jc.ErrorIsNil)
+	err = results.Combine()
+	c.Assert(err, jc.ErrorIsNil)
+	s.st.CheckCalls(c, []testing.StubCall{
+		{"GetRemoteEntity", []interface{}{names.NewModelTag("uuid"), "token-db2:db django:db"}},
+		{"KeyRelation", []interface{}{"db2:db django:db"}},
+		{"GetRemoteEntity", []interface{}{names.NewModelTag("uuid"), "token-db2"}},
+	})
+	ru1.CheckCalls(c, []testing.StubCall{
+		{"InScope", []interface{}{}},
+		{"EnterScope", []interface{}{map[string]interface{}{"foo": "bar"}}},
+	})
+	ru2.CheckCalls(c, []testing.StubCall{
+		{"LeaveScope", []interface{}{}},
+	})
+}
+
+func (s *remoteRelationsSuite) assertRegisterRemoteRelations(c *gc.C) {
+	app := &mockApplication{}
+	app.eps = []state.Endpoint{{
+		ApplicationName: "offeredapp",
+		Relation:        charm.Relation{Name: "local"},
+	}}
+	s.st.applications["offeredapp"] = app
+	s.st.offers = []crossmodel.ApplicationOffer{{
+		OfferName:       "offered",
+		ApplicationName: "offeredapp",
+	}}
+	results, err := s.api.RegisterRemoteRelations(params.RegisterRemoteRelations{
+		Relations: []params.RegisterRemoteRelation{{
+			ApplicationId:     params.RemoteEntityId{ModelUUID: "model-uuid", Token: "app-token"},
+			RelationId:        params.RemoteEntityId{ModelUUID: "model-uuid", Token: "rel-token"},
+			RemoteEndpoint:    params.RemoteEndpoint{Name: "remote"},
+			OfferName:         "offered",
+			LocalEndpointName: "local",
+		}}})
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(results.Results, gc.HasLen, 1)
+	result := results.Results[0]
+	c.Assert(result.Error, gc.IsNil)
+	c.Check(result.Result, jc.DeepEquals, &params.RemoteEntityId{
+		ModelUUID: coretesting.ModelTag.Id(), Token: "token-offeredapp"})
+	expectedRemoteApp := s.st.remoteApplications["remote-apptoken"]
+	expectedRemoteApp.Stub = testing.Stub{} // don't care about api calls
+
+	c.Check(expectedRemoteApp, jc.DeepEquals, &mockRemoteApplication{consumerproxy: true})
+	expectedRel := s.st.relations["offeredapp:local remote-apptoken:remote"]
+	expectedRel.Stub = testing.Stub{} // don't care about api calls
+	c.Check(expectedRel, jc.DeepEquals, &mockRelation{key: "offeredapp:local remote-apptoken:remote"})
+	c.Check(s.st.remoteEntities, gc.HasLen, 2)
+	c.Check(s.st.remoteEntities[names.NewApplicationTag("offeredapp")], gc.Equals, "token-offeredapp")
+	c.Check(s.st.remoteEntities[names.NewRelationTag("offeredapp:local remote-apptoken:remote")], gc.Equals, "rel-token")
+}
+
+func (s *remoteRelationsSuite) TestRegisterRemoteRelations(c *gc.C) {
+	s.assertRegisterRemoteRelations(c)
+}
+
+func (s *remoteRelationsSuite) TestRegisterRemoteRelationsIdempotent(c *gc.C) {
+	s.assertRegisterRemoteRelations(c)
+	s.assertRegisterRemoteRelations(c)
+}
