+++ apiserver/remoterelations/remoterelations_test.go
@@ -15,7 +15,6 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/apiserver/remoterelations"
 	apiservertesting "github.com/juju/juju/apiserver/testing"
-	"github.com/juju/juju/core/crossmodel"
 	"github.com/juju/juju/state"
 	coretesting "github.com/juju/juju/testing"
 )
@@ -43,8 +42,7 @@ func (s *remoteRelationsSuite) SetUpTest(c *gc.C) {
 	}
 
 	s.st = newMockState()
-	pool := &mockStatePool{s.st}
-	api, err := remoterelations.NewRemoteRelationsAPI(s.st, pool, s.resources, s.authorizer)
+	api, err := remoterelations.NewRemoteRelationsAPI(s.st, s.resources, s.authorizer)
 	c.Assert(err, jc.ErrorIsNil)
 	s.api = api
 }
@@ -106,54 +104,6 @@ func (s *remoteRelationsSuite) TestWatchRemoteRelations(c *gc.C) {
 	resource := s.resources.Get("1")
 	c.Assert(resource, gc.NotNil)
 	c.Assert(resource, gc.Implements, new(state.StringsWatcher))
-}
-
-func (s *remoteRelationsSuite) TestPublishLocalRelationsChange(c *gc.C) {
-	s.st.remoteApplications["db2"] = newMockRemoteApplication("db2", "db2url")
-	s.st.remoteEntities[names.NewApplicationTag("db2")] = "token-db2"
-	rel := newMockRelation(1)
-	rel.endpoints = []state.Endpoint{
-		{ApplicationName: "db2"},
-	}
-	ru1 := newMockRelationUnit()
-	ru2 := newMockRelationUnit()
-	rel.units["db2/1"] = ru1
-	rel.units["db2/2"] = ru2
-	s.st.relations["db2:db django:db"] = rel
-	s.st.remoteEntities[names.NewRelationTag("db2:db django:db")] = "token-db2:db django:db"
-	results, err := s.api.PublishLocalRelationChange(params.RemoteRelationsChanges{
-		Changes: []params.RemoteRelationChangeEvent{
-			{
-				Life: params.Alive,
-				ApplicationId: params.RemoteEntityId{
-					ModelUUID: "uuid",
-					Token:     "token-db2"},
-				RelationId: params.RemoteEntityId{
-					ModelUUID: "uuid",
-					Token:     "token-db2:db django:db"},
-				ChangedUnits: []params.RemoteRelationUnitChange{{
-					UnitId:   1,
-					Settings: map[string]interface{}{"foo": "bar"},
-				}},
-				DepartedUnits: []int{2},
-			},
-		},
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	err = results.Combine()
-	c.Assert(err, jc.ErrorIsNil)
-	s.st.CheckCalls(c, []testing.StubCall{
-		{"GetRemoteEntity", []interface{}{names.NewModelTag("uuid"), "token-db2:db django:db"}},
-		{"KeyRelation", []interface{}{"db2:db django:db"}},
-		{"GetRemoteEntity", []interface{}{names.NewModelTag("uuid"), "token-db2"}},
-	})
-	ru1.CheckCalls(c, []testing.StubCall{
-		{"InScope", []interface{}{}},
-		{"EnterScope", []interface{}{map[string]interface{}{"foo": "bar"}}},
-	})
-	ru2.CheckCalls(c, []testing.StubCall{
-		{"LeaveScope", []interface{}{}},
-	})
 }
 
 func (s *remoteRelationsSuite) TestWatchLocalRelationUnits(c *gc.C) {
@@ -389,53 +339,4 @@ func (s *remoteRelationsSuite) TestRelations(c *gc.C) {
 		{"Application", []interface{}{"django"}},
 		{"RemoteApplication", []interface{}{"db2"}},
 	})
-}
-
-func (s *remoteRelationsSuite) assertRegisterRemoteRelations(c *gc.C) {
-	app := newMockApplication("offeredapp")
-	app.eps = []state.Endpoint{{
-		ApplicationName: "offeredapp",
-		Relation:        charm.Relation{Name: "local"},
-	}}
-	s.st.applications["offeredapp"] = app
-	s.st.offers = []crossmodel.ApplicationOffer{{
-		OfferName:       "offered",
-		ApplicationName: "offeredapp",
-	}}
-	results, err := s.api.RegisterRemoteRelations(params.RegisterRemoteRelations{
-		Relations: []params.RegisterRemoteRelation{{
-			ApplicationId:     params.RemoteEntityId{ModelUUID: "model-uuid", Token: "app-token"},
-			RelationId:        params.RemoteEntityId{ModelUUID: "model-uuid", Token: "rel-token"},
-			RemoteEndpoint:    params.RemoteEndpoint{Name: "remote"},
-			OfferName:         "offered",
-			LocalEndpointName: "local",
-		}}})
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(results.Results, gc.HasLen, 1)
-	result := results.Results[0]
-	c.Assert(result.Error, gc.IsNil)
-	c.Check(result.Result, jc.DeepEquals, &params.RemoteEntityId{
-		ModelUUID: coretesting.ModelTag.Id(), Token: "token-offeredapp"})
-	expectedRemoteApp := s.st.remoteApplications["remote-apptoken"]
-	expectedRemoteApp.Stub = testing.Stub{} // don't care about api calls
-	c.Check(expectedRemoteApp, jc.DeepEquals, &mockRemoteApplication{
-		name:          "remote-apptoken",
-		eps:           []charm.Relation{{Name: "remote"}},
-		consumerproxy: true,
-	})
-	expectedRel := s.st.relations["offeredapp:local remote-apptoken:remote"]
-	expectedRel.Stub = testing.Stub{} // don't care about api calls
-	c.Check(expectedRel, jc.DeepEquals, &mockRelation{key: "offeredapp:local remote-apptoken:remote"})
-	c.Check(s.st.remoteEntities, gc.HasLen, 2)
-	c.Check(s.st.remoteEntities[names.NewApplicationTag("offeredapp")], gc.Equals, "token-offeredapp")
-	c.Check(s.st.remoteEntities[names.NewRelationTag("offeredapp:local remote-apptoken:remote")], gc.Equals, "rel-token")
-}
-
-func (s *remoteRelationsSuite) TestRegisterRemoteRelations(c *gc.C) {
-	s.assertRegisterRemoteRelations(c)
-}
-
-func (s *remoteRelationsSuite) TestRegisterRemoteRelationsIdempotent(c *gc.C) {
-	s.assertRegisterRemoteRelations(c)
-	s.assertRegisterRemoteRelations(c)
 }
